added helpful printing function -- also we gotta start dealing with window size

+++ include/tcp/recv_window.h
@@ -21,11 +21,11 @@ void recv_window_chunk_destroy(recv_window_chunk_t* rwc);
 void recv_window_chunk_destroy_total(recv_window_chunk_t* rwc, destructor_f destructor);
 void recv_window_chunk_destroy_free(recv_window_chunk_t* rwc);
 
-recv_window_t recv_window_init(uint32_t window_size, uint32_t ISN);
+recv_window_t recv_window_init(uint16_t window_size, uint32_t ISN);
 void recv_window_destroy(recv_window_t* recv_window);
 recv_window_chunk_t recv_window_get_next(recv_window_t window);
 uint32_t recv_window_get_ack(recv_window_t window);
-uint32_t recv_window_get_size(recv_window_t window);
+uint16_t recv_window_get_size(recv_window_t window);
 void recv_window_receive(recv_window_t window, void* data, uint32_t length, uint32_t seqnum);
 
 #endif // __RECV_WINDOW_H__
